refactor(executor): modularize Codex runtime
