+++ executor/src/agents/codex.rs
@@ -6,7 +6,6 @@ use std::{
     collections::{BTreeMap, HashMap, HashSet},
     env, fs,
     future::Future,
-    io::{self, Write},
     path::{Path, PathBuf},
     pin::Pin,
     process::Stdio,
@@ -27,13 +26,11 @@ use tokio::{
 use crate::{
     agents::runtime_capabilities,
     attachments::{process_prompt, AttachmentPromptProcessor, AttachmentRecord},
-    codex_phase::{codex_phase_is_process, CodexAgentMessagePhaseTracker},
     image_preprocessor::prepare_image_bytes_for_model,
     logging::{log_executor_event, task_fields},
     process_environment,
     protocol::ExecutionRequest,
     runner::{AgentEngine, ExecutionOutcome},
-    runtime_work::codex_stream_debug_enabled,
     server::{
         codex_model_catalog, executor_loopback_base_url,
         local_model_proxy::{self, LocalModelProxyUpstream},
@@ -49,8 +46,6 @@ const DEFAULT_PROVIDER_ID: &str = "wecode-openai";
 pub const CODEX_APP_SERVER_TURN_CANCELLED: &str = "codex app-server turn cancelled";
 const DEFAULT_REASONING_EFFORT: &str = "medium";
 const DEFAULT_NO_PROXY: &str = "localhost,127.0.0.1,::1,host.docker.internal";
-const CODEX_HOME_ENV: &str = "CODEX_HOME";
-const WEGENT_CODEX_HOME_ENV: &str = "WEGENT_CODEX_HOME";
 const CODEX_ROUTER_API_KEY: &str = "wework-local-router";
 const EXECUTOR_INTERNAL_ENV_KEYS: &[&str] = &[
     "WEGENT_EXECUTOR_BINARY",
@@ -68,9 +63,6 @@ const CODEX_APPLY_PATCH_STREAMING_EVENTS_OVERRIDE: &str =
 const CODEX_SUPPRESS_UNSTABLE_FEATURES_WARNING_OVERRIDE: &str =
     "suppress_unstable_features_warning=true";
 const DEFAULT_EXECUTOR_SERVER_PORT: u16 = 10001;
-const CODEX_RAW_LOG_PREVIEW_CHARS: usize = 1200;
-const CODEX_RAW_LOG_LARGE_STRING_CHARS: usize = 2048;
-const CODEX_RAW_LOG_STRING_PREVIEW_CHARS: usize = 240;
 const SIDE_BOUNDARY_PROMPT: &str = r#"Side conversation boundary.
 
 The messages before this boundary are inherited reference context from the main thread.
@@ -93,6 +85,16 @@ const IMAGE_MIME_TYPES: &[&str] = &[
     "image/webp",
     "image/bmp",
 ];
+
+#[path = "codex/diagnostics.rs"]
+mod diagnostics;
+#[path = "codex/home.rs"]
+mod home;
+
+use diagnostics::{json_scalar_field, json_string_field};
+#[cfg(test)]
+use home::WEGENT_CODEX_HOME_ENV;
+use home::{prepare_wework_codex_home, wework_codex_home, CODEX_HOME_ENV};
 
 pub type CodexNotificationSender = mpsc::UnboundedSender<Value>;
 pub type CodexThreadStartedCallback = Box<dyn FnOnce(String) + Send + 'static>;
@@ -183,88 +185,11 @@ pub struct CodexAppServerTurn {
     pub outcome: ExecutionOutcome,
 }
 
-pub type CodexRequestUserInputReceiver = mpsc::Receiver<Value>;
+#[path = "codex/interaction.rs"]
+mod interaction;
 
-#[derive(Default)]
-struct InteractionAnswerRouterState {
-    pending: HashMap<String, oneshot::Sender<Result<Value, String>>>,
-    buffered: HashMap<String, Value>,
-    closed: bool,
-}
-
-struct InteractionAnswerRouter {
-    state: Arc<Mutex<InteractionAnswerRouterState>>,
-}
-
-impl InteractionAnswerRouter {
-    fn new(mut receiver: CodexRequestUserInputReceiver) -> Arc<Self> {
-        let router = Arc::new(Self {
-            state: Arc::new(Mutex::new(InteractionAnswerRouterState::default())),
-        });
-        let state = router.state.clone();
-        tokio::spawn(async move {
-            while let Some(answer) = receiver.recv().await {
-                let mut state = state.lock().await;
-                let key = interaction_answer_key(&answer).or_else(|| {
-                    (state.pending.len() == 1)
-                        .then(|| state.pending.keys().next().cloned())
-                        .flatten()
-                });
-                let Some(key) = key else {
-                    continue;
-                };
-                if let Some(sender) = state.pending.remove(&key) {
-                    let _ = sender.send(Ok(answer));
-                } else {
-                    state.buffered.insert(key, answer);
-                }
-            }
-            let mut state = state.lock().await;
-            state.closed = true;
-            for (_, sender) in state.pending.drain() {
-                let _ = sender.send(Err("request_user_input response channel closed".to_owned()));
-            }
-        });
-        router
-    }
-
-    async fn receive(&self, key: String) -> Result<Value, String> {
-        let receiver = {
-            let mut state = self.state.lock().await;
-            if let Some(answer) = state.buffered.remove(&key) {
-                return Ok(answer);
-            }
-            if state.closed {
-                return Err("request_user_input response router closed".to_owned());
-            }
-            let (sender, receiver) = oneshot::channel();
-            if state.pending.insert(key, sender).is_some() {
-                return Err("duplicate pending interaction correlation key".to_owned());
-            }
-            receiver
-        };
-        receiver
-            .await
-            .map_err(|_| "request_user_input response router closed".to_owned())?
-    }
-}
-
-fn interaction_answer_key(answer: &Value) -> Option<String> {
-    answer
-        .get("requestId")
-        .or_else(|| answer.get("request_id"))
-        .or_else(|| answer.get("itemId"))
-        .or_else(|| answer.get("item_id"))
-        .and_then(interaction_value_key)
-}
-
-fn interaction_value_key(value: &Value) -> Option<String> {
-    match value {
-        Value::String(value) if !value.trim().is_empty() => Some(value.clone()),
-        Value::Number(value) => Some(value.to_string()),
-        _ => None,
-    }
-}
+pub use interaction::CodexRequestUserInputReceiver;
+use interaction::{interaction_value_key, InteractionAnswerRouter};
 
 #[derive(Debug, Clone)]
 pub struct CodexAppServerEngine {
@@ -1786,99 +1711,6 @@ fn spawn_codex_app_server(
         .map_err(|error| format!("failed to start codex app-server: {error}"))
 }
 
-fn wework_codex_home() -> PathBuf {
-    env::var_os(WEGENT_CODEX_HOME_ENV)
-        .filter(|value| !value.is_empty())
-        .map(PathBuf::from)
-        .unwrap_or_else(|| executor_home().join("codex"))
-}
-
-fn prepare_wework_codex_home(codex_home: &Path) -> Result<(), String> {
-    fs::create_dir_all(codex_home).map_err(|error| {
-        format!(
-            "failed to create Codex home {}: {error}",
-            codex_home.display()
-        )
-    })?;
-    link_user_codex_auth(codex_home)?;
-    normalize_wework_codex_config(codex_home)
-}
-
-fn normalize_wework_codex_config(codex_home: &Path) -> Result<(), String> {
-    use toml_edit::{value, DocumentMut};
-
-    let config_path = codex_home.join("config.toml");
-    let content = fs::read_to_string(&config_path).unwrap_or_default();
-    let mut document = content.parse::<DocumentMut>().map_err(|error| {
-        format!(
-            "failed to parse Codex config {}: {error}",
-            config_path.display()
-        )
-    })?;
-    let legacy_instructions = document
-        .get("instructions")
-        .and_then(|item| item.as_str())
-        .unwrap_or_default();
-    let developer_instructions = document
-        .get("developer_instructions")
-        .and_then(|item| item.as_str())
-        .unwrap_or_default();
-    let user_instructions = if legacy_instructions.trim().is_empty() {
-        strip_wework_browser_instructions(developer_instructions).to_owned()
-    } else {
-        legacy_instructions.trim().to_owned()
-    };
-
-    document.remove("instructions");
-    document["developer_instructions"] =
-        value(combined_codex_developer_instructions(&user_instructions));
-    if document
-        .get("personality")
-        .and_then(|item| item.as_str())
-        .is_none()
-    {
-        document["personality"] = value("pragmatic");
-    }
-
-    let next_content = document.to_string();
-    if next_content == content {
-        return Ok(());
-    }
-    let temporary_path = config_path.with_extension("toml.tmp");
-    fs::write(&temporary_path, next_content).map_err(|error| {
-        format!(
-            "failed to write Codex config {}: {error}",
-            temporary_path.display()
-        )
-    })?;
-    if let Ok(metadata) = fs::metadata(&config_path) {
-        fs::set_permissions(&temporary_path, metadata.permissions()).map_err(|error| {
-            format!(
-                "failed to preserve Codex config permissions {}: {error}",
-                temporary_path.display()
-            )
-        })?;
-    }
-    #[cfg(unix)]
-    if !config_path.exists() {
-        use std::os::unix::fs::PermissionsExt;
-        fs::set_permissions(&temporary_path, fs::Permissions::from_mode(0o600)).map_err(
-            |error| {
-                format!(
-                    "failed to secure Codex config permissions {}: {error}",
-                    temporary_path.display()
-                )
-            },
-        )?;
-    }
-    fs::rename(&temporary_path, &config_path).map_err(|error| {
-        format!(
-            "failed to replace Codex config {}: {error}",
-            config_path.display()
-        )
-    })
-}
-
 pub(crate) fn combined_codex_developer_instructions(user_instructions: &str) -> String {
     let user_instructions = user_instructions.trim();
     if user_instructions.is_empty() {
@@ -1892,54 +1724,6 @@ pub(crate) fn strip_wework_browser_instructions(instructions: &str) -> &str {
         .strip_suffix(WEWORK_EMBEDDED_BROWSER_DEVELOPER_INSTRUCTIONS)
         .unwrap_or(instructions)
         .trim()
-}
-
-fn link_user_codex_auth(codex_home: &Path) -> Result<(), String> {
-    let target = codex_home.join("auth.json");
-    if let Ok(metadata) = fs::symlink_metadata(&target) {
-        if metadata.file_type().is_symlink() && !target.exists() {
-            fs::remove_file(&target).map_err(|error| {
-                format!(
-                    "failed to remove stale Codex auth link {}: {error}",
-                    target.display()
-                )
-            })?;
-        } else {
-            return Ok(());
-        }
-    }
-    let Some(source) = user_codex_auth_path().filter(|path| path.is_file()) else {
-        return Ok(());
-    };
-
-    #[cfg(unix)]
-    {
-        std::os::unix::fs::symlink(&source, &target).map_err(|error| {
-            format!(
-                "failed to link Codex auth {} -> {}: {error}",
-                target.display(),
-                source.display()
-            )
-        })
-    }
-    #[cfg(not(unix))]
-    {
-        fs::copy(&source, &target).map(|_| ()).map_err(|error| {
-            format!(
-                "failed to copy Codex auth {} -> {}: {error}",
-                source.display(),
-                target.display()
-            )
-        })
-    }
-}
-
-fn user_codex_auth_path() -> Option<PathBuf> {
-    env::var_os(CODEX_HOME_ENV)
-        .filter(|value| !value.is_empty())
-        .map(PathBuf::from)
-        .map(|home| home.join("auth.json"))
-        .or_else(|| dirs::home_dir().map(|home| home.join(".codex").join("auth.json")))
 }
 
 #[cfg(test)]
@@ -2257,541 +2041,10 @@ impl JsonRpcConnection {
     }
 }
 
-#[derive(Default)]
-struct CodexRunState {
-    final_text: String,
-    saw_delta: bool,
-    agent_message_phases: CodexAgentMessagePhaseTracker,
-    root_thread_id: Option<String>,
-    goal_status: Option<String>,
-}
+#[path = "codex/run_state.rs"]
+mod run_state;
 
-impl CodexRunState {
-    fn set_root_thread_id(&mut self, thread_id: impl Into<String>) {
-        self.root_thread_id = Some(thread_id.into());
-    }
-
-    fn set_goal_status(&mut self, status: impl Into<String>) {
-        self.goal_status = Some(status.into().to_ascii_lowercase());
-    }
-
-    fn goal_is_active(&self) -> bool {
-        self.goal_status
-            .as_deref()
-            .is_some_and(|status| status.eq_ignore_ascii_case("active"))
-    }
-
-    fn reset_turn_output(&mut self) {
-        self.final_text.clear();
-        self.saw_delta = false;
-        self.agent_message_phases = CodexAgentMessagePhaseTracker::default();
-    }
-
-    fn handle_message(&mut self, message: &Value) -> Option<ExecutionOutcome> {
-        match message.get("method").and_then(Value::as_str) {
-            Some("thread/started") => {
-                if self.root_thread_id.is_none() {
-                    if let Some(thread_id) = stream_thread_id(message_params(message)) {
-                        self.root_thread_id = Some(thread_id);
-                    }
-                }
-                None
-            }
-            Some("turn/started") => {
-                if !self.is_subagent_message(message_params(message)) {
-                    self.reset_turn_output();
-                }
-                None
-            }
-            Some("thread/goal/updated") => {
-                if let Some(status) = message_params(message)
-                    .get("goal")
-                    .and_then(|goal| goal.get("status"))
-                    .and_then(Value::as_str)
-                {
-                    self.set_goal_status(status);
-                }
-                None
-            }
-            Some("thread/goal/cleared") => {
-                self.goal_status = None;
-                None
-            }
-            Some("item/started") => {
-                if self.is_subagent_message(message_params(message)) {
-                    return None;
-                }
-                self.agent_message_phases
-                    .observe_item(message_params(message));
-                None
-            }
-            Some("item/agentMessage/delta") => {
-                if self.is_subagent_message(message_params(message)) {
-                    return None;
-                }
-                self.append_delta(message_params(message));
-                None
-            }
-            Some("item/completed") => {
-                let params = message_params(message);
-                if self.is_subagent_message(params) {
-                    return None;
-                }
-                self.append_completed_message(params);
-                self.agent_message_phases.forget_item(params);
-                None
-            }
-            Some("turn/completed")
-                if !self.is_subagent_message(message_params(message))
-                    && is_root_codex_turn_event(message_params(message)) =>
-            {
-                Some(self.completed(message_params(message)))
-            }
-            Some("turn/completed") => None,
-            Some("error") => {
-                let params = message_params(message);
-                log_codex_run_state_error(params);
-                if codex_error_will_retry(params) {
-                    return None;
-                }
-                Some(ExecutionOutcome::Failed {
-                    message: codex_error_message(params),
-                })
-            }
-            _ => None,
-        }
-    }
-
-    fn is_subagent_message(&self, params: &Value) -> bool {
-        codex_agent_path(params)
-            .or_else(|| params.get("item").and_then(codex_agent_path))
-            .is_some_and(|agent_path| agent_path != "/root")
-            || self
-                .root_thread_id
-                .as_deref()
-                .is_some_and(|root_thread_id| {
-                    stream_thread_id(params).is_some_and(|thread_id| thread_id != root_thread_id)
-                })
-    }
-
-    fn append_delta(&mut self, params: &Value) {
-        let text = params.get("delta").and_then(Value::as_str).unwrap_or("");
-        let phase = self.agent_message_phases.phase_for_delta(params);
-        if codex_phase_is_process(phase.as_deref()) {
-            log_codex_run_state_text(
-                "delta",
-                "skip_process",
-                phase.as_deref(),
-                params,
-                params,
-                text,
-            );
-            return;
-        }
-        if let Some(delta) = params.get("delta").and_then(Value::as_str) {
-            log_codex_run_state_text(
-                "delta",
-                "append_final",
-                phase.as_deref(),
-                params,
-                params,
-                delta,
-            );
-            self.final_text.push_str(delta);
-            self.saw_delta = true;
-        }
-    }
-
-    fn append_completed_message(&mut self, params: &Value) {
-        let phase = self.agent_message_phases.phase_for_item(params);
-        if self.saw_delta {
-            log_codex_run_state_text(
-                "completed",
-                "skip_after_delta",
-                phase.as_deref(),
-                params,
-                params,
-                "",
-            );
-            return;
-        }
-        let item = params.get("item").unwrap_or(params);
-        let item_type = item
-            .get("type")
-            .and_then(Value::as_str)
-            .unwrap_or("")
-            .replace('_', "")
-            .to_ascii_lowercase();
-        if item_type == "plan" {
-            let text = extract_text(item).unwrap_or_default();
-            log_codex_run_state_text(
-                "completed",
-                "skip_plan",
-                phase.as_deref(),
-                params,
-                item,
-                &text,
-            );
-            return;
-        }
-        if !matches!(item_type.as_str(), "agentmessage" | "message") {
-            log_codex_run_state_text("completed", "skip_non_message", None, params, item, "");
-            return;
-        }
-        if codex_phase_is_process(phase.as_deref()) {
-            let text = extract_text(item).unwrap_or_default();
-            log_codex_run_state_text(
-                "completed",
-                "skip_process",
-                phase.as_deref(),
-                params,
-                item,
-                &text,
-            );
-            return;
-        }
-        if item
-            .get("role")
-            .and_then(Value::as_str)
-            .is_some_and(|role| role != "assistant")
-        {
-            log_codex_run_state_text(
-                "completed",
-                "skip_non_assistant",
-                phase.as_deref(),
-                params,
-                item,
-                "",
-            );
-            return;
-        }
-        if let Some(text) = extract_text(item) {
-            log_codex_run_state_text(
-                "completed",
-                "set_final",
-                phase.as_deref(),
-                params,
-                item,
-                &text,
-            );
-            self.final_text = text;
-            self.saw_delta = true;
-        }
-    }
-
-    fn completed(&self, params: &Value) -> ExecutionOutcome {
-        let status = params
-            .get("turn")
-            .and_then(|turn| turn.get("status"))
-            .or_else(|| params.get("status"))
-            .and_then(Value::as_str)
-            .unwrap_or("completed")
-            .to_ascii_lowercase();
-        match status.as_str() {
-            "completed" | "complete" | "succeeded" => ExecutionOutcome::Completed {
-                content: self.final_text.clone(),
-            },
-            "cancelled" | "canceled" | "interrupted" => {
-                ExecutionOutcome::Cancelled { message: status }
-            }
-            other => ExecutionOutcome::Failed {
-                message: format!("codex turn ended with status {other}"),
-            },
-        }
-    }
-}
-
-fn codex_error_will_retry(params: &Value) -> bool {
-    params
-        .get("willRetry")
-        .or_else(|| params.get("will_retry"))
-        .and_then(Value::as_bool)
-        .unwrap_or(false)
-}
-
-fn is_root_codex_turn_event(params: &Value) -> bool {
-    let turn = params.get("turn").unwrap_or(params);
-    codex_agent_path(turn)
-        .or_else(|| codex_agent_path(params))
-        .map_or(true, |agent_path| agent_path == "/root")
-}
-
-fn stream_thread_id(value: &Value) -> Option<String> {
-    value
-        .get("threadId")
-        .or_else(|| value.get("thread_id"))
-        .and_then(Value::as_str)
-        .or_else(|| {
-            value
-                .get("thread")
-                .and_then(|thread| thread.get("id"))
-                .and_then(Value::as_str)
-        })
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-}
-
-fn codex_agent_path(value: &Value) -> Option<String> {
-    value
-        .get("agent_path")
-        .or_else(|| value.get("agentPath"))
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-}
-
-fn log_codex_run_state_text(
-    source: &str,
-    action: &str,
-    resolved_phase: Option<&str>,
-    params: &Value,
-    item: &Value,
-    text: &str,
-) {
-    if source == "delta" && !codex_stream_debug_enabled() {
-        return;
-    }
-
-    log_executor_event(
-        "codex run state text classification",
-        &[
-            ("source", source.to_owned()),
-            ("action", action.to_owned()),
-            (
-                "resolved_phase",
-                resolved_phase.unwrap_or("<none>").to_owned(),
-            ),
-            ("item_id", json_string_field(params, "itemId")),
-            ("params_type", json_string_field(params, "type")),
-            ("params_phase", json_string_field(params, "phase")),
-            ("params_channel", json_string_field(params, "channel")),
-            ("item_type", json_string_field(item, "type")),
-            ("item_phase", json_string_field(item, "phase")),
-            ("item_channel", json_string_field(item, "channel")),
-            (
-                "payload_type",
-                nested_json_string_field(item, "payload", "type"),
-            ),
-            (
-                "payload_phase",
-                nested_json_string_field(item, "payload", "phase"),
-            ),
-            (
-                "payload_channel",
-                nested_json_string_field(item, "payload", "channel"),
-            ),
-            ("text_len", text.len().to_string()),
-            ("text_preview", truncate_log_text(text, 160)),
-        ],
-    );
-}
-
-fn log_codex_run_state_error(params: &Value) {
-    let message = codex_error_message(params);
-    let params_json = serde_json::to_string(params)
-        .unwrap_or_else(|error| format!("failed to serialize codex error params: {error}"));
-    log_executor_event(
-        "codex run state error",
-        &[
-            ("message", message),
-            ("code", json_string_field(params, "code")),
-            ("params_len", params_json.len().to_string()),
-            ("params_preview", truncate_log_text(&params_json, 500)),
-        ],
-    );
-}
-
-fn log_codex_raw_turn_message(message: &Value) {
-    let method = message.get("method").and_then(Value::as_str).unwrap_or("");
-    if matches!(
-        method,
-        "item/agentMessage/delta" | "item/reasoning/delta" | "item/reasoningSummary/delta"
-    ) && !codex_stream_debug_enabled()
-    {
-        return;
-    }
-
-    if !matches!(
-        method,
-        "item/agentMessage/delta"
-            | "item/reasoning/delta"
-            | "item/reasoningSummary/delta"
-            | "item/fileChange/patchUpdated"
-            | "turn/plan/updated"
-            | "item/started"
-            | "item/completed"
-            | "turn/completed"
-            | "error"
-    ) {
-        return;
-    }
-
-    let params = message_params(message);
-    let item = params.get("item").unwrap_or(params);
-    let raw_len = serialized_json_len(message)
-        .map(|length| length.to_string())
-        .unwrap_or_else(|error| format!("failed to measure codex raw message: {error}"));
-    let raw_preview = codex_raw_log_preview(message);
-    log_executor_event(
-        "codex raw turn message",
-        &[
-            ("method", method.to_owned()),
-            ("message_id", json_string_field(message, "id")),
-            ("params_keys", json_object_keys(params)),
-            ("params_type", json_string_field(params, "type")),
-            ("params_phase", json_string_field(params, "phase")),
-            ("params_channel", json_string_field(params, "channel")),
-            ("params_item_id", json_string_field(params, "item_id")),
-            ("params_message_id", json_string_field(params, "message_id")),
-            (
-                "params_output_index",
-                json_scalar_field(params, "output_index"),
-            ),
-            (
-                "params_content_index",
-                json_scalar_field(params, "content_index"),
-            ),
-            ("item_keys", json_object_keys(item)),
-            ("item_type", json_string_field(item, "type")),
-            ("item_id", json_string_field(item, "id")),
-            ("item_phase", json_string_field(item, "phase")),
-            ("item_channel", json_string_field(item, "channel")),
-            (
-                "item_turn_id",
-                nested_json_string_field(
-                    item,
-                    "internal_chat_message_metadata_passthrough",
-                    "turn_id",
-                ),
-            ),
-            ("raw_len", raw_len),
-            ("raw_preview", raw_preview),
-        ],
-    );
-}
-
-struct ByteCounter {
-    length: usize,
-}
-
-impl Write for ByteCounter {
-    fn write(&mut self, buffer: &[u8]) -> io::Result<usize> {
-        self.length += buffer.len();
-        Ok(buffer.len())
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        Ok(())
-    }
-}
-
-fn serialized_json_len(value: &Value) -> serde_json::Result<usize> {
-    let mut counter = ByteCounter { length: 0 };
-    serde_json::to_writer(&mut counter, value)?;
-    Ok(counter.length)
-}
-
-fn codex_raw_log_preview(value: &Value) -> String {
-    let sanitized = sanitize_codex_raw_log_value(value, None);
-    let preview = serde_json::to_string(&sanitized)
-        .unwrap_or_else(|error| format!("failed to serialize codex raw message preview: {error}"));
-    truncate_log_text(&preview, CODEX_RAW_LOG_PREVIEW_CHARS)
-}
-
-fn sanitize_codex_raw_log_value(value: &Value, key: Option<&str>) -> Value {
-    match value {
-        Value::Object(object) => Value::Object(
-            object
-                .iter()
-                .map(|(key, value)| {
-                    (
-                        key.clone(),
-                        sanitize_codex_raw_log_value(value, Some(key.as_str())),
-                    )
-                })
-                .collect(),
-        ),
-        Value::Array(items) => Value::Array(
-            items
-                .iter()
-                .map(|item| sanitize_codex_raw_log_value(item, None))
-                .collect(),
-        ),
-        Value::String(text) if should_summarize_codex_raw_log_string(key, text) => {
-            Value::String(format!(
-                "[{} chars omitted; preview: {}]",
-                text.chars().count(),
-                truncate_log_text(text, CODEX_RAW_LOG_STRING_PREVIEW_CHARS)
-            ))
-        }
-        _ => value.clone(),
-    }
-}
-
-fn should_summarize_codex_raw_log_string(key: Option<&str>, text: &str) -> bool {
-    matches!(
-        key,
-        Some("aggregatedOutput")
-            | Some("toolOutput")
-            | Some("tool_output")
-            | Some("toolOutputDelta")
-            | Some("tool_output_delta")
-            | Some("output")
-            | Some("stdout")
-            | Some("stderr")
-    ) || text.len() > CODEX_RAW_LOG_LARGE_STRING_CHARS
-}
-
-fn json_string_field(value: &Value, key: &str) -> String {
-    value
-        .get(key)
-        .and_then(Value::as_str)
-        .unwrap_or("")
-        .to_owned()
-}
-
-fn json_scalar_field(value: &Value, key: &str) -> String {
-    value
-        .get(key)
-        .map(|value| {
-            value
-                .as_str()
-                .map(str::to_owned)
-                .unwrap_or_else(|| value.to_string())
-        })
-        .unwrap_or_default()
-}
-
-fn json_object_keys(value: &Value) -> String {
-    value
-        .as_object()
-        .map(|object| object.keys().cloned().collect::<Vec<_>>().join(","))
-        .unwrap_or_default()
-}
-
-fn nested_json_string_field(value: &Value, object_key: &str, key: &str) -> String {
-    value
-        .get(object_key)
-        .and_then(|object| object.get(key))
-        .and_then(Value::as_str)
-        .unwrap_or("")
-        .to_owned()
-}
-
-fn truncate_log_text(text: &str, max_chars: usize) -> String {
-    let mut result = String::new();
-    for (index, ch) in text.chars().enumerate() {
-        if index >= max_chars {
-            result.push('…');
-            return result;
-        }
-        result.push(ch);
-    }
-    result
-}
-
+use run_state::{log_codex_raw_turn_message, stream_thread_id, CodexRunState};
 fn initialize_params() -> Value {
     json!({
         "clientInfo": {
@@ -5024,1436 +4277,5 @@ fn extract_text(item: &Value) -> Option<String> {
 }
 
 #[cfg(test)]
-mod tests {
-    use serde_json::json;
-
-    use super::*;
-
-    #[tokio::test]
-    async fn interaction_answer_router_matches_reverse_order_answers() {
-        let (sender, receiver) = mpsc::channel(2);
-        let router = InteractionAnswerRouter::new(receiver);
-        let first = {
-            let router = router.clone();
-            tokio::spawn(async move { router.receive("41".to_owned()).await })
-        };
-        let second = {
-            let router = router.clone();
-            tokio::spawn(async move { router.receive("42".to_owned()).await })
-        };
-
-        sender
-            .send(json!({"requestId": 42, "answers": {"choice": "second"}}))
-            .await
-            .expect("second answer should be sent");
-        sender
-            .send(json!({"requestId": 41, "answers": {"choice": "first"}}))
-            .await
-            .expect("first answer should be sent");
-
-        assert_eq!(
-            first.await.expect("first waiter should join").unwrap()["answers"]["choice"],
-            "first"
-        );
-        assert_eq!(
-            second.await.expect("second waiter should join").unwrap()["answers"]["choice"],
-            "second"
-        );
-    }
-
-    #[tokio::test]
-    async fn interaction_answer_router_rejects_waiters_after_channel_closes() {
-        let (sender, receiver) = mpsc::channel(1);
-        let router = InteractionAnswerRouter::new(receiver);
-        drop(sender);
-        tokio::time::timeout(Duration::from_secs(1), async {
-            loop {
-                if router.state.lock().await.closed {
-                    break;
-                }
-                tokio::task::yield_now().await;
-            }
-        })
-        .await
-        .expect("router should observe the closed response channel");
-
-        let result = tokio::time::timeout(
-            Duration::from_secs(1),
-            router.receive("closed-request".to_owned()),
-        )
-        .await
-        .expect("closed router should not leave the waiter pending");
-
-        assert_eq!(
-            result.unwrap_err(),
-            "request_user_input response router closed"
-        );
-    }
-
-    #[test]
-    fn normalize_reasoning_effort_preserves_supported_codex_levels() {
-        for effort in ["low", "medium", "high", "xhigh", "max", "ultra"] {
-            assert_eq!(normalize_reasoning_effort(Some(effort)), effort);
-        }
-    }
-
-    #[test]
-    fn normalize_reasoning_effort_maps_aliases_to_supported_codex_levels() {
-        for (value, expected) in [
-            ("minimal", "low"),
-            ("轻度", "low"),
-            ("中等", "medium"),
-            ("extra high", "xhigh"),
-            ("x-high", "xhigh"),
-            ("最高", "max"),
-            ("maximum", "max"),
-            ("极高", "ultra"),
-        ] {
-            assert_eq!(normalize_reasoning_effort(Some(value)), expected);
-        }
-    }
-
-    #[test]
-    fn normalize_reasoning_effort_uses_default_for_disabled_or_unknown_values() {
-        for value in [None, Some("off"), Some("unknown")] {
-            assert_eq!(normalize_reasoning_effort(value), DEFAULT_REASONING_EFFORT);
-        }
-    }
-
-    #[test]
-    fn wework_codex_home_defaults_to_executor_home_codex() {
-        let _lock = crate::test_env::lock();
-        let home = unique_test_path("wework-codex-home-default");
-        let _executor_home = EnvRestore::capture("WEGENT_EXECUTOR_HOME");
-        let _wework_codex_home = EnvRestore::capture(WEGENT_CODEX_HOME_ENV);
-        let _codex_home = EnvRestore::capture(CODEX_HOME_ENV);
-
-        env::set_var("WEGENT_EXECUTOR_HOME", &home);
-        env::remove_var(WEGENT_CODEX_HOME_ENV);
-        env::set_var(
-            CODEX_HOME_ENV,
-            home.join("user-codex-should-not-be-wework-home"),
-        );
-
-        assert_eq!(wework_codex_home(), home.join("codex"));
-
-        let _ = fs::remove_dir_all(home);
-    }
-
-    #[test]
-    fn wework_codex_home_prefers_explicit_wework_home() {
-        let _lock = crate::test_env::lock();
-        let executor_home = unique_test_path("wework-codex-home-executor");
-        let codex_home = unique_test_path("wework-codex-home-explicit");
-        let _executor_home = EnvRestore::capture("WEGENT_EXECUTOR_HOME");
-        let _wework_codex_home = EnvRestore::capture(WEGENT_CODEX_HOME_ENV);
-        let _codex_home = EnvRestore::capture(CODEX_HOME_ENV);
-
-        env::set_var("WEGENT_EXECUTOR_HOME", &executor_home);
-        env::set_var(WEGENT_CODEX_HOME_ENV, &codex_home);
-        env::set_var(CODEX_HOME_ENV, executor_home.join("ignored-codex"));
-
-        assert_eq!(wework_codex_home(), codex_home);
-
-        let _ = fs::remove_dir_all(executor_home);
-        let _ = fs::remove_dir_all(codex_home);
-    }
-
-    #[test]
-    fn prepare_wework_codex_home_links_user_auth() {
-        let _lock = crate::test_env::lock();
-        let root = unique_test_path("wework-codex-home-auth");
-        let user_codex_home = root.join("user-codex");
-        let codex_home = root.join("wework-codex");
-        let source_auth = user_codex_home.join("auth.json");
-        let _codex_home = EnvRestore::capture(CODEX_HOME_ENV);
-
-        fs::create_dir_all(source_auth.parent().expect("auth parent should exist"))
-            .expect("user Codex home should be created");
-        fs::write(&source_auth, br#"{"token":"shared"}"#).expect("auth should be written");
-        env::set_var(CODEX_HOME_ENV, &user_codex_home);
-
-        prepare_wework_codex_home(&codex_home).expect("Codex home should be prepared");
-
-        let linked_auth = codex_home.join("auth.json");
-        assert!(linked_auth.is_file());
-        #[cfg(unix)]
-        assert_eq!(
-            fs::read_link(&linked_auth).expect("auth should be a symlink"),
-            source_auth
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn prepare_wework_codex_home_replaces_stale_auth_link() {
-        let _lock = crate::test_env::lock();
-        let root = unique_test_path("wework-codex-home-stale-auth");
-        let user_codex_home = root.join("user-codex");
-        let codex_home = root.join("wework-codex");
-        let source_auth = user_codex_home.join("auth.json");
-        let stale_source = root.join("missing-auth.json");
-        let linked_auth = codex_home.join("auth.json");
-        let _codex_home = EnvRestore::capture(CODEX_HOME_ENV);
-
-        fs::create_dir_all(source_auth.parent().expect("auth parent should exist"))
-            .expect("user Codex home should be created");
-        fs::create_dir_all(&codex_home).expect("WeWork Codex home should be created");
-        fs::write(&source_auth, br#"{"token":"shared"}"#).expect("auth should be written");
-        std::os::unix::fs::symlink(&stale_source, &linked_auth)
-            .expect("stale auth link should be created");
-        env::set_var(CODEX_HOME_ENV, &user_codex_home);
-
-        prepare_wework_codex_home(&codex_home).expect("Codex home should be prepared");
-
-        assert_eq!(
-            fs::read_link(&linked_auth).expect("auth should be a symlink"),
-            source_auth
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn prepare_wework_codex_home_migrates_base_instruction_override() {
-        let _lock = crate::test_env::lock();
-        let root = unique_test_path("wework-codex-config-migration");
-        let codex_home = root.join("codex");
-        fs::create_dir_all(&codex_home).expect("Codex home should be created");
-        fs::write(
-            codex_home.join("config.toml"),
-            "instructions = \"用中文回复\"\n",
-        )
-        .expect("legacy config should be written");
-
-        prepare_wework_codex_home(&codex_home).expect("Codex config should be normalized");
-
-        let config = fs::read_to_string(codex_home.join("config.toml"))
-            .expect("normalized config should be readable");
-        assert!(!config
-            .lines()
-            .any(|line| line.starts_with("instructions =")));
-        assert!(config.contains("developer_instructions"));
-        assert!(config.contains("用中文回复"));
-        assert!(config.contains("browser_navigate"));
-        assert!(config.contains("personality = \"pragmatic\""));
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn codex_raw_log_preview_summarizes_large_command_output() {
-        let output = "x".repeat(4096);
-        let message = json!({
-            "method": "item/completed",
-            "params": {
-                "item": {
-                    "id": "cmd-1",
-                    "type": "commandExecution",
-                    "aggregatedOutput": output,
-                }
-            }
-        });
-
-        let preview = codex_raw_log_preview(&message);
-
-        assert!(preview.contains("4096 chars omitted"));
-        assert!(!preview.contains(&"x".repeat(512)));
-        assert_eq!(
-            serialized_json_len(&message).expect("message length should serialize"),
-            serde_json::to_string(&message)
-                .expect("message should serialize")
-                .len()
-        );
-    }
-
-    #[test]
-    fn codex_launch_config_enables_streaming_patch_updates() {
-        let request = ExecutionRequest {
-            prompt: Value::String("create a file".to_owned()),
-            model_config: json!({
-                "model_id": "gpt-5.5-codex",
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        let launch_config = build_codex_launch_config(&request);
-
-        assert!(launch_config
-            .config_overrides
-            .contains(&CODEX_APPLY_PATCH_STREAMING_EVENTS_OVERRIDE.to_owned()));
-        assert!(launch_config
-            .config_overrides
-            .contains(&CODEX_SUPPRESS_UNSTABLE_FEATURES_WARNING_OVERRIDE.to_owned()));
-    }
-
-    #[test]
-    fn custom_model_without_catalog_entry_uses_upstream_id() {
-        let request = ExecutionRequest {
-            model_config: json!({
-            "model_id": "kimi-for-coding",
-            "tool_profile": "function",
-            "codex_responses_compat_proxy": true
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        assert_eq!(
-            codex_request_model(&request).as_deref(),
-            Some("kimi-for-coding")
-        );
-    }
-
-    #[test]
-    fn kimi_k3_profile_uses_the_built_in_catalog_entry() {
-        let request = ExecutionRequest {
-            model_config: json!({
-                "model_id": "kimi-k3",
-                "tool_profile": "function",
-                "codex_catalog_model_id": codex_model_catalog::KIMI_K3_MODEL,
-                "codex_responses_compat_proxy": true
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        assert_eq!(
-            codex_request_model(&request).as_deref(),
-            Some(codex_model_catalog::KIMI_K3_MODEL)
-        );
-    }
-
-    #[test]
-    fn cloud_model_uses_provider_model_id_for_catalog_capabilities() {
-        let request = ExecutionRequest {
-            model_config: json!({
-                "model_id": "openai-gpt-5.6-luna(海外)",
-                "codex_catalog_model_id": "gpt-5.6-luna",
-                "codex_responses_compat_proxy": true
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        assert_eq!(
-            codex_request_model(&request).as_deref(),
-            Some("gpt-5.6-luna")
-        );
-    }
-
-    #[test]
-    fn shell_profile_uses_explicit_catalog_capabilities() {
-        let request = ExecutionRequest {
-            model_config: json!({
-                "model_id": "openai-gpt-5.6-luna(海外)",
-                "codex_catalog_model_id": "gpt-5.6-luna",
-                "tool_profile": "shell",
-                "codex_responses_compat_proxy": true
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        assert_eq!(
-            codex_request_model(&request).as_deref(),
-            Some("gpt-5.6-luna")
-        );
-    }
-
-    #[test]
-    fn custom_shell_profile_without_catalog_entry_uses_upstream_id() {
-        let request = ExecutionRequest {
-            model_config: json!({
-                "model_id": "native-model",
-                "tool_profile": "shell",
-                "codex_responses_compat_proxy": true
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        assert_eq!(
-            codex_request_model(&request).as_deref(),
-            Some("native-model")
-        );
-    }
-
-    #[test]
-    fn internal_catalog_provider_is_never_used_for_thread_inference() {
-        let request = ExecutionRequest {
-            model_config: json!({
-                "model_id": "gpt-5.4",
-                "model_provider": "wework-catalog"
-            }),
-            ..ExecutionRequest::default()
-        };
-        let launch_config = build_codex_launch_config(&request);
-
-        assert_eq!(launch_config.model_provider.as_deref(), Some("openai"));
-        for params in [
-            thread_start_params(&request, &launch_config),
-            thread_resume_params("thread-1", &request, &launch_config),
-            thread_fork_params("thread-1", None, &request, &launch_config),
-        ] {
-            assert_eq!(params["modelProvider"], "openai");
-        }
-    }
-
-    #[test]
-    fn configured_inference_provider_reads_the_unmodified_user_config() {
-        let root = unique_test_path("configured-inference-provider");
-        fs::create_dir_all(&root).expect("test directory should be created");
-        let config_path = root.join("config.toml");
-        fs::write(
-            &config_path,
-            "model_provider = \"wework-e2e\"\n[model_providers.wework-e2e]\nbase_url = \"http://127.0.0.1/v1\"\n",
-        )
-        .expect("config should be written");
-
-        assert_eq!(
-            configured_inference_model_provider_from_path(&config_path),
-            "wework-e2e"
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn user_configured_provider_routes_inference_through_the_local_router() {
-        let _lock = crate::test_env::lock();
-        let root = unique_test_path("configured-provider-router");
-        let _wework_codex_home = EnvRestore::capture(WEGENT_CODEX_HOME_ENV);
-        let _api_key = EnvRestore::capture("WEWORK_TEST_MODEL_API_KEY");
-        fs::create_dir_all(&root).expect("test directory should be created");
-        fs::write(
-            root.join("config.toml"),
-            "model_provider = \"wework-e2e\"\n[model_providers.wework-e2e]\nbase_url = \"http://127.0.0.1:3456/v1\"\nenv_key = \"WEWORK_TEST_MODEL_API_KEY\"\nwire_api = \"responses\"\n",
-        )
-        .expect("config should be written");
-        env::set_var(WEGENT_CODEX_HOME_ENV, &root);
-        env::set_var("WEWORK_TEST_MODEL_API_KEY", "test-key");
-        let request = ExecutionRequest {
-            model_config: json!({
-                "model_id": "gpt-test",
-                "runtime_config": {
-                    "codex": {
-                        "use_user_config": true,
-                        "configured": true
-                    }
-                }
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        let launch_config = build_codex_launch_config(&request);
-
-        assert_eq!(
-            launch_config.model_provider.as_deref(),
-            Some(codex_model_catalog::PROVIDER_ID)
-        );
-        assert!(launch_config.local_proxy_registration.is_some());
-        assert!(launch_config.config_overrides.iter().any(|value| {
-            value.starts_with("model_providers.wework-router.base_url=\"http://127.0.0.1:")
-                && value.contains("/v1/codex-router/model-")
-        }));
-        for params in [
-            thread_start_params(&request, &launch_config),
-            thread_resume_params("thread-1", &request, &launch_config),
-            thread_fork_params("thread-1", None, &request, &launch_config),
-        ] {
-            assert_eq!(params["modelProvider"], codex_model_catalog::PROVIDER_ID);
-        }
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn configured_inference_provider_rejects_the_internal_catalog_provider() {
-        let root = unique_test_path("configured-catalog-provider");
-        fs::create_dir_all(&root).expect("test directory should be created");
-        let config_path = root.join("config.toml");
-        fs::write(&config_path, "model_provider = \"wework-catalog\"\n")
-            .expect("config should be written");
-
-        assert_eq!(
-            configured_inference_model_provider_from_path(&config_path),
-            "openai"
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn codex_launch_config_forwards_web_search_mode() {
-        let request = ExecutionRequest {
-            prompt: Value::String("create a file".to_owned()),
-            model_config: json!({
-                "model_id": "gpt-5.5-codex",
-                "web_search": "disabled",
-                "image_generation": false,
-                "model_context_window": 128000,
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        let launch_config = build_codex_launch_config(&request);
-        let params = thread_start_params(&request, &launch_config);
-        let config = params
-            .get("config")
-            .and_then(Value::as_object)
-            .expect("thread config should be present");
-
-        assert_eq!(config.get("web_search"), Some(&json!("disabled")));
-        assert_eq!(config.get("features.image_generation"), Some(&json!(false)));
-        assert_eq!(config.get("model_context_window"), Some(&json!(128000)));
-    }
-
-    #[test]
-    fn codex_launch_config_routes_marked_responses_models_through_compat_proxy() {
-        let request = ExecutionRequest {
-            prompt: Value::String("create a file".to_owned()),
-            model_config: json!({
-                "model_id": "mimo-v2.5-pro",
-                "base_url": "http://models.local/v1",
-                "api_key": "sk-local",
-                "api_format": "responses",
-                "codex_responses_compat_proxy": true,
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        let launch_config = build_codex_launch_config(&request);
-
-        assert_eq!(
-            launch_config.model_provider.as_deref(),
-            Some(codex_model_catalog::PROVIDER_ID)
-        );
-        assert!(launch_config.config_overrides.iter().any(|override_value| {
-            override_value.starts_with("model_providers.wework-router.base_url=\"http://127.0.0.1:")
-                && override_value.contains("/v1/codex-router/model-")
-        }));
-        assert!(!launch_config
-            .config_overrides
-            .iter()
-            .any(|override_value| override_value.contains("experimental_bearer_token")));
-    }
-
-    #[test]
-    fn codex_launch_config_forwards_runtime_proxy_env() {
-        let request = ExecutionRequest {
-            prompt: Value::String("create a file".to_owned()),
-            model_config: json!({
-                "model_id": "gpt-5.5-codex",
-                "proxy": {
-                    "url": "http://127.0.0.1:7890"
-                },
-                "runtime_config": {
-                    "codex": {
-                        "use_proxy": true
-                    }
-                }
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        let launch_config = build_codex_launch_config(&request);
-
-        assert_eq!(
-            launch_config.env.get("HTTP_PROXY").map(String::as_str),
-            Some("http://127.0.0.1:7890")
-        );
-        assert_eq!(
-            launch_config.env.get("HTTPS_PROXY").map(String::as_str),
-            Some("http://127.0.0.1:7890")
-        );
-        assert_eq!(
-            launch_config.env.get("ALL_PROXY").map(String::as_str),
-            Some("http://127.0.0.1:7890")
-        );
-    }
-
-    #[test]
-    fn codex_launch_config_does_not_forward_task_identity() {
-        let request = ExecutionRequest {
-            task_id: "task-525".to_owned(),
-            auth_token: Some("task-jwt".to_owned()),
-            skill_identity_token: Some("skill-jwt".to_owned()),
-            user_name: Some("alice".to_owned()),
-            prompt: Value::String("create a file".to_owned()),
-            model_config: json!({
-                "model_id": "gpt-5.5-codex",
-            }),
-            ..ExecutionRequest::default()
-        };
-
-        let launch_config = build_codex_launch_config(&request);
-        let params = thread_start_params(&request, &launch_config);
-        let config = params
-            .get("config")
-            .and_then(Value::as_object)
-            .expect("thread config should include shell env");
-
-        assert!(!launch_config.env.contains_key("WEGENT_TASK_ID"));
-        assert!(!launch_config.env.contains_key("AUTH_TOKEN"));
-        assert!(config
-            .get("shell_environment_policy.set.WEGENT_TASK_ID")
-            .is_none());
-        assert!(config
-            .get("shell_environment_policy.set.AUTH_TOKEN")
-            .is_none());
-        assert!(config
-            .get("shell_environment_policy.set.WEGENT_SKILL_IDENTITY_TOKEN")
-            .is_none());
-        assert!(config
-            .get("shell_environment_policy.set.WEGENT_SKILL_USER_NAME")
-            .is_none());
-    }
-
-    #[test]
-    fn persistent_codex_app_server_launch_config_keeps_only_process_settings() {
-        let request_launch_config = CodexLaunchConfig {
-            env: BTreeMap::from([("HTTP_PROXY".to_owned(), "http://127.0.0.1:7890".to_owned())]),
-            config_overrides: vec![
-                "model_provider=wecode-openai".to_owned(),
-                "model_catalog_json=\"/tmp/wework-models.json\"".to_owned(),
-                "mcp_servers.wework.command=\"node\"".to_owned(),
-            ],
-            model_provider: Some("wecode-openai".to_owned()),
-            effort: Some("high".to_owned()),
-            summary: Some("auto".to_owned()),
-            ..CodexLaunchConfig::default()
-        };
-
-        let launch_config = persistent_codex_app_server_launch_config(&request_launch_config);
-
-        assert_eq!(
-            launch_config.env.get("HTTP_PROXY").map(String::as_str),
-            Some("http://127.0.0.1:7890")
-        );
-        assert!(launch_config
-            .config_overrides
-            .iter()
-            .any(|value| value == "model_provider=wework-router"));
-        assert!(!launch_config
-            .config_overrides
-            .iter()
-            .any(|value| value.starts_with("model_catalog_json=")));
-        assert!(launch_config
-            .config_overrides
-            .contains(&"goals=true".to_owned()));
-        assert!(launch_config.model_provider.is_none());
-        assert!(launch_config.effort.is_none());
-        assert!(launch_config.summary.is_none());
-    }
-
-    #[test]
-    fn persistent_process_does_not_inherit_request_model_overrides() {
-        assert_eq!(
-            persistent_codex_app_server_launch_config(&CodexLaunchConfig::default())
-                .config_overrides,
-            persistent_codex_app_server_launch_config(&CodexLaunchConfig {
-                config_overrides: vec!["model=gpt-custom".to_owned()],
-                ..CodexLaunchConfig::default()
-            })
-            .config_overrides
-        );
-    }
-
-    #[test]
-    fn codex_run_state_keeps_commentary_agent_delta_out_of_final_content() {
-        let mut state = CodexRunState::default();
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "phase": "commentary",
-                    "delta": "I will inspect."
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .expect("turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: String::new()
-            }
-        );
-    }
-
-    #[test]
-    fn goal_created_during_turn_keeps_notification_reader_alive() {
-        let mut state = CodexRunState::default();
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "thread/goal/updated",
-                "params": {
-                    "threadId": "thread-1",
-                    "goal": { "status": "active" }
-                }
-            }))
-            .is_none());
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "threadId": "thread-1",
-                    "turn": { "status": "completed" }
-                }
-            }))
-            .expect("turn completion should produce an outcome");
-
-        assert!(should_wait_for_goal_continuation(&outcome, &state));
-    }
-
-    #[test]
-    fn codex_run_state_keeps_commentary_channel_delta_out_of_final_content() {
-        let mut state = CodexRunState::default();
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "channel": "commentary",
-                    "delta": "I will inspect."
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .expect("turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: String::new()
-            }
-        );
-    }
-
-    #[test]
-    fn codex_run_state_keeps_completed_plan_out_of_final_content() {
-        let mut state = CodexRunState::default();
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/completed",
-                "params": {
-                    "item": {
-                        "id": "turn-1-plan",
-                        "type": "plan",
-                        "text": "# Plan\n\n- Execute the steps."
-                    }
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .expect("turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: String::new()
-            }
-        );
-    }
-
-    #[test]
-    fn codex_run_state_routes_item_id_deltas_by_started_phase() {
-        let mut state = CodexRunState::default();
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/started",
-                "params": {
-                    "item": {
-                        "id": "msg-commentary",
-                        "type": "agentMessage",
-                        "phase": "commentary",
-                        "text": ""
-                    }
-                }
-            }))
-            .is_none());
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "itemId": "msg-commentary",
-                    "delta": "I will inspect."
-                }
-            }))
-            .is_none());
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/started",
-                "params": {
-                    "item": {
-                        "id": "msg-final",
-                        "type": "agentMessage",
-                        "phase": "final_answer",
-                        "text": ""
-                    }
-                }
-            }))
-            .is_none());
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "itemId": "msg-final",
-                    "delta": "Done."
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .expect("turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: "Done.".to_owned()
-            }
-        );
-    }
-
-    #[test]
-    fn codex_run_state_keeps_unphased_agent_delta_as_final_content() {
-        let mut state = CodexRunState::default();
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "delta": "Current directory: /tmp/project"
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .expect("turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: "Current directory: /tmp/project".to_owned()
-            }
-        );
-    }
-
-    #[test]
-    fn turn_start_params_includes_plan_collaboration_mode_when_requested() {
-        let mut request = ExecutionRequest {
-            prompt: Value::String("plan this".to_owned()),
-            model_config: json!({
-                "model_id": "gpt-5.5",
-            }),
-            ..ExecutionRequest::default()
-        };
-        request.extra.insert(
-            "collaborationMode".to_owned(),
-            Value::String("plan".to_owned()),
-        );
-        let launch_config = CodexLaunchConfig {
-            effort: Some("high".to_owned()),
-            ..CodexLaunchConfig::default()
-        };
-
-        let params = turn_start_params(
-            "thread-1",
-            &request,
-            &launch_config,
-            vec![json!({"type": "text", "text": "plan this"})],
-        );
-
-        assert_eq!(params["collaborationMode"]["mode"], "plan");
-        assert_eq!(params["collaborationMode"]["settings"]["model"], "gpt-5.5");
-        assert_eq!(
-            params["collaborationMode"]["settings"]["reasoningEffort"],
-            "high"
-        );
-        assert!(params["collaborationMode"]["settings"]["developerInstructions"].is_null());
-    }
-
-    #[test]
-    fn turn_start_params_includes_client_user_message_id() {
-        let mut request = ExecutionRequest::default();
-        request.extra.insert(
-            "client_user_message_id".to_owned(),
-            Value::String("runtime-local-pane-1".to_owned()),
-        );
-
-        let params = turn_start_params(
-            "thread-1",
-            &request,
-            &CodexLaunchConfig::default(),
-            Vec::new(),
-        );
-
-        assert_eq!(params["clientUserMessageId"], "runtime-local-pane-1");
-    }
-
-    #[test]
-    fn codex_permission_profile_is_applied_to_thread_and_turn_requests() {
-        let request = ExecutionRequest::default();
-        let launch_config = CodexLaunchConfig::default();
-        let thread_start = thread_start_params(&request, &launch_config);
-        let thread_resume = thread_resume_params("thread-1", &request, &launch_config);
-        let thread_fork = thread_fork_params("thread-1", None, &request, &launch_config);
-        let turn_start = turn_start_params("thread-1", &request, &launch_config, Vec::new());
-
-        for params in [thread_start, thread_resume, thread_fork, turn_start] {
-            assert_eq!(
-                params["permissions"],
-                CODEX_DANGER_FULL_ACCESS_PERMISSION_PROFILE
-            );
-            assert!(params.get("sandboxPolicy").is_none());
-            assert!(params.get("sandbox").is_none());
-        }
-    }
-
-    #[test]
-    fn codex_permission_profile_validation_rejects_effective_downgrade() {
-        let response = json!({
-            "activePermissionProfile": {"id": ":workspace"},
-            "sandbox": {"type": "workspaceWrite", "networkAccess": false},
-        });
-
-        let error = validate_codex_permission_profile("thread/resume", &response)
-            .expect_err("workspace-write must not be accepted");
-
-        assert!(error.contains("active_profile=:workspace"));
-        assert!(error.contains("sandbox=workspaceWrite"));
-    }
-
-    #[test]
-    fn codex_permission_profile_validation_accepts_effective_full_access() {
-        let response = json!({
-            "activePermissionProfile": {"id": ":danger-full-access"},
-            "sandbox": {"type": "dangerFullAccess"},
-        });
-
-        validate_codex_permission_profile("thread/resume", &response).unwrap();
-    }
-
-    #[test]
-    fn turn_input_expands_absolute_skill_markdown_mentions_for_app_server() {
-        let input = turn_input(&Value::String(
-            "[$linear](/Users/me/.codex/plugins/linear/skills/linear/SKILL.md) triage".to_owned(),
-        ));
-
-        assert_eq!(
-            input,
-            vec![
-                json!({"type": "text", "text": "$linear triage", "text_elements": []}),
-                json!({
-                    "type": "skill",
-                    "name": "linear",
-                    "path": "/Users/me/.codex/plugins/linear/skills/linear/SKILL.md",
-                }),
-            ]
-        );
-    }
-
-    #[test]
-    fn turn_input_expands_legacy_skill_markdown_mentions_for_app_server() {
-        let input = turn_input(&Value::String(
-            "[$linear](skill:///Users/me/.codex/plugins/linear/skills/linear/SKILL.md) triage"
-                .to_owned(),
-        ));
-
-        assert_eq!(
-            input,
-            vec![
-                json!({"type": "text", "text": "$linear triage", "text_elements": []}),
-                json!({
-                    "type": "skill",
-                    "name": "linear",
-                    "path": "/Users/me/.codex/plugins/linear/skills/linear/SKILL.md",
-                }),
-            ]
-        );
-    }
-
-    #[test]
-    fn turn_input_deduplicates_legacy_and_absolute_references_to_the_same_skill() {
-        let input = turn_input(&Value::String(
-            "[$linear](skill:///Users/me/skills/linear/SKILL.md) then [$linear](/Users/me/skills/linear/SKILL.md)"
-                .to_owned(),
-        ));
-
-        assert_eq!(
-            input,
-            vec![
-                json!({
-                    "type": "text",
-                    "text": "$linear then $linear",
-                    "text_elements": [],
-                }),
-                json!({
-                    "type": "skill",
-                    "name": "linear",
-                    "path": "/Users/me/skills/linear/SKILL.md",
-                }),
-            ]
-        );
-    }
-
-    #[test]
-    fn turn_input_expands_app_and_plugin_markdown_mentions_for_app_server() {
-        let input = turn_input(&Value::String(
-            "Use [$calendar](app://google-calendar) and [$sample](plugin://sample@test)".to_owned(),
-        ));
-
-        assert_eq!(
-            input,
-            vec![
-                json!({
-                    "type": "text",
-                    "text": "Use $calendar and @sample",
-                    "text_elements": [],
-                }),
-                json!({
-                    "type": "mention",
-                    "name": "calendar",
-                    "path": "app://google-calendar",
-                }),
-                json!({
-                    "type": "mention",
-                    "name": "sample",
-                    "path": "plugin://sample@test",
-                }),
-            ]
-        );
-    }
-
-    #[test]
-    fn turn_input_converts_composer_file_references_to_plain_paths() {
-        let input = turn_input(&Value::String(
-            "Inspect [$frontend](folder://%2FUsers%2Fme%2FMy%20Project%2Ffrontend) and [$auth.ts](file://%2FUsers%2Fme%2FMy%20Project%2Ffrontend%2Fauth.ts)"
-                .to_owned(),
-        ));
-
-        assert_eq!(
-            input,
-            vec![json!({
-                "type": "text",
-                "text": "Inspect \"/Users/me/My Project/frontend\" and \"/Users/me/My Project/frontend/auth.ts\"",
-                "text_elements": [],
-            })]
-        );
-    }
-
-    #[test]
-    fn codex_launch_config_includes_cdp_browser_mcp_server() {
-        let _lock = crate::test_env::lock();
-        let home = env::temp_dir().join(format!("codex-browser-mcp-{}", std::process::id()));
-        let old_home = env::var_os("WEGENT_EXECUTOR_HOME");
-        let old_bridge_addr = env::var_os(WEWORK_EMBEDDED_BROWSER_BRIDGE_ADDR_ENV);
-        env::set_var("WEGENT_EXECUTOR_HOME", &home);
-        env::set_var(WEWORK_EMBEDDED_BROWSER_BRIDGE_ADDR_ENV, "127.0.0.1:43127");
-        let request = ExecutionRequest {
-            task_id: "task:123".to_owned(),
-            ..ExecutionRequest::default()
-        };
-
-        let launch_config = build_codex_launch_config(&request);
-        let params = thread_start_params(&request, &launch_config);
-        let config = params
-            .get("config")
-            .and_then(Value::as_object)
-            .expect("thread config should be present");
-        assert!(!config.contains_key("developer_instructions"));
-        assert_eq!(
-            config["skills.config"],
-            json!([
-                {
-                    "name": "browser:control-in-app-browser",
-                    "enabled": false,
-                },
-                {
-                    "name": "chrome:control-chrome",
-                    "enabled": false,
-                },
-            ])
-        );
-        assert_eq!(config["features.non_prefixed_mcp_tool_names"], true);
-        assert_eq!(
-            config["features.code_mode.direct_only_tool_namespaces"],
-            json!([WEWORK_BROWSER_MCP_SERVER_NAME])
-        );
-        assert_eq!(
-            config["mcp_servers.wework_browser.command"],
-            env::current_exe().unwrap().display().to_string()
-        );
-        assert_eq!(
-            config["mcp_servers.wework_browser.args"],
-            json!(["browser-mcp-server"])
-        );
-        assert_eq!(config["mcp_servers.wework_browser.startup_timeout_sec"], 15);
-        assert_eq!(config["mcp_servers.wework_browser.tool_timeout_sec"], 60);
-        assert_eq!(
-            config["mcp_servers.wework_browser.default_tools_approval_mode"],
-            "approve"
-        );
-        assert_eq!(
-            config["mcp_servers.wework_browser.env.WEWORK_EMBEDDED_BROWSER_BRIDGE_URL"],
-            "http://127.0.0.1:43127"
-        );
-        assert_eq!(
-            config["mcp_servers.wework_browser.env.WEWORK_EMBEDDED_BROWSER_LABEL"],
-            "workspace-browser-task-123"
-        );
-
-        if let Some(old_home) = old_home {
-            env::set_var("WEGENT_EXECUTOR_HOME", old_home);
-        } else {
-            env::remove_var("WEGENT_EXECUTOR_HOME");
-        }
-        if let Some(old_bridge_addr) = old_bridge_addr {
-            env::set_var(WEWORK_EMBEDDED_BROWSER_BRIDGE_ADDR_ENV, old_bridge_addr);
-        } else {
-            env::remove_var(WEWORK_EMBEDDED_BROWSER_BRIDGE_ADDR_ENV);
-        }
-    }
-
-    #[test]
-    fn turn_start_params_includes_default_collaboration_mode_when_requested() {
-        let mut request = ExecutionRequest {
-            prompt: Value::String("continue this".to_owned()),
-            model_config: json!({
-                "model_id": "gpt-5.5",
-            }),
-            ..ExecutionRequest::default()
-        };
-        request.extra.insert(
-            "collaborationMode".to_owned(),
-            Value::String("default".to_owned()),
-        );
-        let launch_config = CodexLaunchConfig {
-            effort: Some("medium".to_owned()),
-            ..CodexLaunchConfig::default()
-        };
-
-        let params = turn_start_params(
-            "thread-1",
-            &request,
-            &launch_config,
-            vec![json!({"type": "text", "text": "continue this"})],
-        );
-
-        assert_eq!(params["collaborationMode"]["mode"], "default");
-        assert_eq!(params["collaborationMode"]["settings"]["model"], "gpt-5.5");
-        assert_eq!(
-            params["collaborationMode"]["settings"]["reasoningEffort"],
-            "medium"
-        );
-        assert!(params["collaborationMode"]["settings"]["developerInstructions"].is_null());
-    }
-
-    #[test]
-    fn thread_goal_set_params_maps_initial_goal() {
-        let params = thread_goal_set_params(
-            "thread-1",
-            &json!({
-                "objective": "ship the feature",
-                "status": "paused",
-                "tokenBudget": 1200,
-            }),
-        )
-        .expect("initial goal should map to Codex goal params");
-
-        assert_eq!(
-            params,
-            json!({
-                "threadId": "thread-1",
-                "objective": "ship the feature",
-                "status": "paused",
-                "tokenBudget": 1200,
-            })
-        );
-    }
-
-    #[test]
-    fn thread_goal_set_params_rejects_empty_objective() {
-        let error = thread_goal_set_params("thread-1", &json!({"objective": "   "}))
-            .expect_err("empty objective should be rejected");
-
-        assert_eq!(error, "initial goal objective is required");
-    }
-
-    #[test]
-    fn active_root_turn_notification_uses_item_turn_id_and_ignores_completed_or_child_turns() {
-        let mut state = CodexRunState::default();
-        state.set_root_thread_id("thread-root");
-
-        let active_item = json!({
-            "method": "item/started",
-            "params": {
-                "threadId": "thread-root",
-                "turnId": "turn-current",
-                "item": { "type": "reasoning" }
-            }
-        });
-        assert_eq!(
-            active_root_turn_notification_id(&active_item, &state).as_deref(),
-            Some("turn-current")
-        );
-
-        let completed_turn = json!({
-            "method": "turn/completed",
-            "params": {
-                "threadId": "thread-root",
-                "turn": { "id": "turn-current", "status": "completed" }
-            }
-        });
-        assert_eq!(
-            active_root_turn_notification_id(&completed_turn, &state),
-            None
-        );
-
-        let child_item = json!({
-            "method": "item/started",
-            "params": {
-                "threadId": "thread-root",
-                "turnId": "child-turn",
-                "agentPath": "/root/worker",
-                "item": { "type": "reasoning" }
-            }
-        });
-        assert_eq!(active_root_turn_notification_id(&child_item, &state), None);
-    }
-
-    #[test]
-    fn initial_progress_excludes_user_echo_retry_errors_and_subagents() {
-        let mut state = CodexRunState::default();
-        state.set_root_thread_id("thread-root");
-
-        let user_echo = json!({
-            "method": "item/completed",
-            "params": {
-                "threadId": "thread-root",
-                "item": { "type": "userMessage", "text": "hello" }
-            }
-        });
-        assert!(!codex_notification_has_initial_progress(&user_echo, &state));
-
-        let retryable_error = json!({
-            "method": "error",
-            "params": {
-                "threadId": "thread-root",
-                "message": "Reconnecting... 1/5",
-                "willRetry": true
-            }
-        });
-        assert!(!codex_notification_has_initial_progress(
-            &retryable_error,
-            &state
-        ));
-
-        let subagent_tool = json!({
-            "method": "item/started",
-            "params": {
-                "threadId": "thread-root",
-                "agentPath": "/root/worker",
-                "item": { "type": "commandExecution" }
-            }
-        });
-        assert!(!codex_notification_has_initial_progress(
-            &subagent_tool,
-            &state
-        ));
-
-        let root_tool = json!({
-            "method": "item/started",
-            "params": {
-                "threadId": "thread-root",
-                "item": { "type": "commandExecution" }
-            }
-        });
-        assert!(codex_notification_has_initial_progress(&root_tool, &state));
-    }
-
-    #[test]
-    fn codex_run_state_ignores_subagent_turn_completion() {
-        let mut state = CodexRunState::default();
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "delta": "Still working"
-                }
-            }))
-            .is_none());
-        assert!(state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "turn": {
-                        "status": "completed",
-                        "agent_path": "/root/worker"
-                    }
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "turn": {
-                        "status": "completed",
-                        "agent_path": "/root"
-                    }
-                }
-            }))
-            .expect("root turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: "Still working".to_owned()
-            }
-        );
-    }
-
-    #[test]
-    fn codex_run_state_ignores_cross_thread_final_deltas() {
-        let mut state = CodexRunState::default();
-        state.set_root_thread_id("root-thread");
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "threadId": "child-thread",
-                    "turnId": "child-turn",
-                    "itemId": "msg-child",
-                    "delta": "child"
-                }
-            }))
-            .is_none());
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "threadId": "root-thread",
-                    "turnId": "root-turn",
-                    "itemId": "msg-root",
-                    "delta": "root"
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "threadId": "root-thread",
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .expect("root turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: "root".to_owned()
-            }
-        );
-    }
-
-    #[test]
-    fn codex_run_state_ignores_cross_thread_turn_completion() {
-        let mut state = CodexRunState::default();
-        state.set_root_thread_id("root-thread");
-
-        assert!(state
-            .handle_message(&json!({
-                "method": "item/agentMessage/delta",
-                "params": {
-                    "threadId": "root-thread",
-                    "turnId": "root-turn",
-                    "itemId": "msg-root",
-                    "delta": "root"
-                }
-            }))
-            .is_none());
-        assert!(state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "threadId": "child-thread",
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .is_none());
-
-        let outcome = state
-            .handle_message(&json!({
-                "method": "turn/completed",
-                "params": {
-                    "threadId": "root-thread",
-                    "turn": {
-                        "status": "completed"
-                    }
-                }
-            }))
-            .expect("root turn completion should produce an outcome");
-
-        assert_eq!(
-            outcome,
-            ExecutionOutcome::Completed {
-                content: "root".to_owned()
-            }
-        );
-    }
-}
+#[path = "codex/tests.rs"]
+mod tests;
